backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
drf-nested-routers==0.94.2
setuptools
django-filter==25.1
django-extensions==3.2.1
pillow==11.1.0